fix(webhooks): remove system event retrigger file
